warnings: avoid deprecation warning from fmt
`fmt` has deprecated implicit `enum` -> `int` conversions by its
interfaces, so just do the explicit conversion here instead.

+++ packages/seacas/libraries/ioss/src/Ioss_Region.C
@@ -1655,7 +1655,7 @@ namespace Ioss {
                "\n\nERROR: The entity named '{}' of type {} which is being aliased to '{}' does "
                "not exist in "
                "region '{}'.\n",
-               db_name, type, alias, name());
+               db_name, static_cast<int>(type), alias, name());
     IOSS_ERROR(errmsg);
   }
